requirements.txt
Flask
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-Script
sqlalchemy-utils
sqlalchemy
alembic

psycopg2-binary
psycopg2

schema

PyJWT

celery

gunicorn
python-dotenv